Rename Branches to TransfersControlFlow in SideEffects
`SideEffects::Branches` used to mean only branches and returns, but
after #2873 it now includes possibly throwing expressions, such as
`throw`s or `call`s (only when EH feature is enabled). While I'm not
sure if it was intended, now this effect name is a misnomer, so this
patch renames it to `SideEffects::TransfersControlFlow`.

This also fixes a few existing errors in sideffects.js test.

But I'm not sure what users of binaryen.js expects from this property.
Do users of bianryen.js want this property to include possibly throwing
expressions? We also have `SideEffects::Throws` for its own sake. If
users don't want throwing expressions to be included in
`SideEffects::Branches`, it would be better to change it back to not
include `throws` and leave name as is.

cc @dcodeIO