Default mainReadsParams to true in standalone mode
The process of DCE'ing the argument handling is already handled in
a different way in standalone more.   In standalone mode the entry
point is `_start` which takes no args, and argv code is included
on-demand via the presence or absence the wasi syscalls for argument
processing (__wasi_args_get/__wasi_args_sizes_get).