Move Stack IR out of default passes
This avoids Stack IR from being generated and thrown away in the
middle of an optimization pipeline. This change is partially
preparation for a future in which tuple lowering is performed as a
part of lowering to Stack IR or something similar. In that case, the
lowering would add locals and trying to throw away the Stack IR (or
similar construct) would become an error.