Update flags used in generate_lld_tests
The `--no-gc-sections` was added as part of #2857 but is not
needed and in fact changes the output of some tests.

`--experimental-pic` is needed these days when building shared
libraries with emscripten's abi.

After these two changes I verfied that the following command
generated no local changes (i.e. is a no-op):

  ./scripts/test/generate_lld_tests.py --binaryen-bin=$PWD/../binaryen-out/bin/ $PWD/../llvm-build/bin/ $PWD/../emscripten