Implement changes in reference types proposal
This updates reference types proposal up-to-date, per
WebAssembly/reference-types#87.

Only maybe half of tests have been updated, so CI will not pass. I was
in the process of updating them fully, but then I noticed #3084, so I'm
not sure if I should continue to do this anymore. I asked him if he
wanted to do it himself to prevent this situation but apparently my
questions weren't answered and I didn't know that he was working on this
at the same time.

I'm uploading this anyway in case this is necessary for discussions.