Add a C source for for test/lld/em_asm_O0 (#3045)
That had just a wat file, with no C. This adds a C file as best I
can guess - looks pretty close - and updates all the lld tests with
scripts/test/generate_lld_tests.py and ./auto_update_tests.py lld

As the diff shows, the handwritten wat was very different than what
emcc+lld emit now. I think we must have switches EM_ASMs to
be variadic at some point? That is what they currently are, and
would explain the diff.

See the discussion that led to this in #3044