Improve inlining "heavyweight" (#3085)
Split that mode into an option to check for loops (which indicate a function
is "heavy") and a constant check for having calls. The case of calls is
different as we would need more logic to avoid infinite recursion if we are
willing to inling functions with calls.

Practically, this renames allowHeavyweight to allowFunctionsWithLoops.