Fix wasm2js memory import in case it is minified (#3113)
It was hardcoded as "env.memory", which is usually correct. But if we minify
import names, as in -O3 in emscripten, we need to use the minified name.

Note how in the test it now emits

var memory = env.a;
for the import.

Fixes emscripten-core/emscripten#12123

This was not noticed earlier since that import is only used in memory
growth. The tests that would catch it are wasm2js3.test*memory_growth*
but we only run wasm2js1 on CI. I'll add testing after this lands.