DWARF: Fix abbreviation lookups, they are relative to 1 (#3158)
Apparently I misunderstood the DWARF spec on this. Abbreviation offsets are all
relative to 1 (as 0 is not a valid number). For some reason I thought the first DIE's
index was the "base", as in practice LLVM always emits the lowest index there,
and that's what the LLVM YAML code suggested to me.