Assign import names consistently between text and binaryn reader (#3238)
The s-parser was assigning numbers names per-type where as
the binaryn reader was using the global import count as the
number to append.

This change switches to use per-element count which I think
it preferable as it increases the stability of the auto-generated
names.  e.g. memory is now always named `$mimport0`.