Fuzzer: Handle log_execution instrumentation, and warn on unknown imports (#3271)
log_execution can be useful to run on a fuzz testcase for debugging purposes. This
adds support to --fuzz-exec for printing out those values.

IOW, this PR allows the following debugging flow: you run

  wasm-opt --log-execution

on the wasm, then run it in the fuzzer,

  wasm-opt --fuzz-exec[-before]

and you get that logging printed out (showing where we enter functions, exit them,
etc.).

Also, in general --fuzz-exec ignores unknown imports (so that it can be run on more wasm
files, for at least some fuzz testing there). This also adds a warning in that case, so it
is less surprising (the behavior does not change in this PR).